Uploaded Rscript for RF GWAS model
I am updating a workflow from my dissertation research to run a random forest model in an epiGWAS with cytosine methylation sites as markers. At the time, cytosine methylation was of particular interest for its potential utility for plant breeding. During my PhD I tackled this using a joint linkage mapping approach with a custom function in R, so this will be interesting to see how advancements in technology have enabled this research